pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "MIMIC-IV-MEDS"
dynamic = ["version"]
authors = [
  { name="Matthew McDermott", email="mattmcdermott8@gmail.com" },
]
description = "An ETL pipeline to extract MIMIC-IV data into the MEDS format."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "polars~=1.30.0",
  "meds-transforms~=0.5.3",
  "meds-extract==0.3.0",
  "requests",
  "beautifulsoup4",
  "hydra-core",
  "hydra-joblib-launcher",
]

[tool.setuptools_scm]

[project.optional-dependencies]
dev = ["pre-commit<4"]
tests = ["pytest", "pytest-cov"]
local_parallelism = ["hydra-joblib-launcher"]
slurm_parallelism = ["hydra-submitit-launcher"]

[project.scripts]
MEDS_extract-MIMIC_IV = "MIMIC_IV_MEDS.__main__:main"

[project.urls]
Homepage = "https://github.com/Medical-Event-Data-Standard/MIMIC_IV_MEDS"
Issues = "https://github.com/Medical-Event-Data-Standard/MIMIC_IV_MEDS/issues"
